Make `has_reified_type_constraint` consistent with `remove_erased_generics`
Summary:
I believe there exists an inefficiency in how we emit `VerifyParamType` and `VerifyParamTypeTS` as well as `VerifyRetTypeTS` and `VerifyRetType`. This diff seeks to correct that inefficiency

## Context:

Currently, when we compile a function, we emit type checks for the parameter types and the return types. For instance, given the following function:

```
function foo(A $a): B { ... }
```

We emit a runtime type check for the function parameter `$a` to check that it is indeed the type `A` and we emit a runtime type check for the return value to ensure that it is of type `B`.

The bytecode we emit for the runtime type check differs if the type in question (either the parameter type or the return type) has generics or not.

If the type in question has generics, we may need to emit `VerifyParamTypeTS` and `VerifyRetTypeTS`. This bytecode is more expensive and is necessary because the class in question may have reified generics. However, there are certain circumstances where we do not need to emit the more expensive bytecode and can emit the cheaper `VerifyParamType` and `VerifyRetType`. We can do this when we know that the type in question only has erased generics.

So in the following example:

```
function foo<T>(A<T> $a): B<int> { ... }
```

We should be able to emit the cheaper `VerifyParamType` instead of `VerifyParamTypeTS` because we know that the generic `T` is an erased function type parameter. However, for the return type, we would still want to emit the `VerifyRetTypeTS`, because we can not infer that `B` has only erased generics.

Currently, there exists logic that attempts to determine whether we should emit `VerifyParamType` and `VerifyRetType` instead of the more expensive `TS` versions in the fact or specified generics. That code is in:

https://www.internalfb.com/code/fbsource/[bf5cec5cee41d614c6354c89c9bf742338d7bfb4]/fbcode/hphp/hack/src/hhbc/hhbc_by_ref/emit_body.rs?lines=600-614

For the function parameter type check and in

https://www.internalfb.com/code/fbsource/[bf5cec5cee41d614c6354c89c9bf742338d7bfb4]/fbcode/hphp/hack/src/hhbc/hhbc_by_ref/try_finally_rewriter.rs?lines=155-159

For the return type check.

## Inefficiency

However, I believe there is an inefficiency in this logic. In both places, we perform a `remove_erased_generics` and then a `has_reified_type_constraint` call to determine whether a type has erased generics.

The first call to `remove_erased_generics` actually modifies the hint to produce a new hint. It does the following:

https://www.internalfb.com/code/fbsource/[bf5cec5cee41d614c6354c89c9bf742338d7bfb4]/fbcode/hphp/hack/src/hhbc/hhbc_by_ref/reified_generics_helpers.rs?lines=197-203

It replaces any erased type parameter with an underscore. So essentially, in the example:

```
function foo<T>(A<T> $a): B<int> { ... }
```

It transforms the type `A<T>` into `A<_>` because the `T` is an erased type parameter.

Then in `has_reified_type_constraint` we check to see whether the hint has any reified type constraints. How does it do this check?

https://www.internalfb.com/code/fbsource/[bf5cec5cee41d614c6354c89c9bf742338d7bfb4]/fbcode/hphp/hack/src/hhbc/hhbc_by_ref/reified_generics_helpers.rs?lines=54-62

It just checks the names to see if they are in the erased tparam list. Unfortunately, `_` is not in the erased type parameter list. Only `T` is. As a result, `has_reified_type_constraint` would believe that `A<_>` has a reified generic! Therefore we would emit the more expensive `VerifyParamTypeTS` check instead of emitting the cheaper `VerifyParamType` check.

## Fix

The simplest fix here is to do a check to see if the type parameter is `_`. That is what this diff currently does. There are alternatives, but may be less convincing.

```
        h_iter.all(|h| {
            if let Hint_::Happly(Id(_, ref id), ref apply_hints) = *h.1 {
                if apply_hints.is_empty() {
                    return id == "_" || erased_tparams.contains(id);
                }
            }
            false
        })
```

Reviewed By: oulgen

Differential Revision: D32216898

fbshipit-source-id: 16f650145c3bf5ca4cc26309204712c385cef319