Move partial integration tests to strict
Summary:
See the title. I updated the partial test cases to be strict, mostly by adding type annotations.

For the global inference tests, instead of adding annotations, we add FIXMEs. Adding type annotations would defeat the purpose of global inference tests (that is to add the missing annotations).

One test that significantly changed behaviour is about reporting the type coverage of a file. Now that we are eliminating partial, there's no point of this. So I changed the file to have type annotations hence the expected test result changed from 0% to 100%. See `test_lsp.py`.

Reviewed By: andrewjkennedy

Differential Revision: D32136101

fbshipit-source-id: 17dec029158f250136457fb14d17f9c2e8e6c6a7