Backout D30226954 in favor of another mitigation strategy
Summary: Backout of {D30226954 (https://github.com/facebook/hhvm/commit/51baa6362401f5d1ef5f28a1ba72889078cebb1c)}. The `custom_rule` in that diff reduced the frequency of the issue, but didn't eliminate it. The next diff in this stack will attempt to address T96769599 using a filegroup (as suggested in Buck Q&A).

Reviewed By: ljw1004

Differential Revision: D32269033

fbshipit-source-id: daec5718459e4b865860055bc01b4df6ab51cd44