Add ParsedFile types for output of direct decl parser
Summary: The output of the direct decl parser is often coupled with a few other things. We return the mode of the file we parsed, and for OCaml consumers, we also return the hash of the file and the hash of each decl in the file. This has led to unwieldy tuple types like `decls * FileInfo.mode option * Int64.t option * Int64.t list option`. Instead, define `ParsedFile` and `ParsedFileWithHashes` records to make these types clearer. This also makes it easier to add more information about the file in the future (e.g., file attributes, as in D32014540).

Reviewed By: bobrenjc93

Differential Revision: D32074153

fbshipit-source-id: b85444202e465e28fbb4a028f0eababd2f61d52a