Support SDT on lambdas
Summary:
Support the use of `<<__SupportDynamicType>>` on lambdas.
* If a lambda is explicitly marked with the SDT attribute, then it undergoes the same checks as a top-level SDT function.
* It is assumed to be implicitly marked if the enclosing function, class, or method is marked SDT.
* The resulting function type is an SDT function.

The `function_dynamically_callable` function is moved from `typing_toplevel.ml` to `typing.ml` so that the code for checking lambdas can make use of it.

8 more tests are now preserved by pessimisation after this diff.

TODO: clean up the code that extracts (optional) decl types from hints. Currently this ends up merging global inference results, which is incorrect for use in checking enforceability for SDT.

Differential Revision: D32250145

fbshipit-source-id: e69b2694bd7c83fc9abb7c9fc390b5dee3f34dc5