Remove --disable-modes option to hh_single_type_check
Summary:
`--disable-modes` is used to treat partial files as strict ones. In preparation to removing partial mode altogether, we get rid of this soon-to-be defunct flag.

As a result of removing this option from the parser, all files with `// partial` mode are treated as strict unconditionally. See the changes in `aast_parser.rs` and `rust_parser_ffi.rs` to see how `Mpartial` is converted to `Mstrict` without `disable_modes`.

Reviewed By: andrewjkennedy

Differential Revision: D32069294

fbshipit-source-id: 8e01b263eb304787671fd80eba41c5c57e8ddc78