Delete partial mode
Summary:
This diff deletes `Mpartial` mode and the APIs that make use of it.

Most notably, the pragmas at the beginning of the files such as `// partial`, `// blah`, no longer have additional meaning and are just comments. We previously threw parsing errors when we encounter something other than `// strict` or `// partial`.

Reviewed By: andrewjkennedy

Differential Revision: D32062815

fbshipit-source-id: 3fe66fc9c66840cfa8caf05e6ac62581176de54d