Create WatchmanWatcherTest
Summary:
Facts wants to know if a given Watchman result is "fresh" or "incremental". If it's a fresh update, then every file in the update is a file that exists in the repo. If it's an incremental update, then every file in the update is a *changed* file since our last update.

If we ask Watchman for the files that have changed since a given point in time, and Watchman returns `"is_fresh_instance": true`, then we have a fresh update and should treat it accordingly.

If we ask Watchman for the files that have changed since a given *commit*, our update is incremental by our definition. But Watchman will return `"is_fresh_instance": true` in its results, because that's not Watchman's definition of freshness.

We dropped this subtlety while refactoring `WatchmanWatcher`, leading to On Demand WWW failures that D32097190 (https://github.com/facebook/hhvm/commit/89c590d114c9f62bc374422e3663a1cf4dd75c6b) ultimately fixed.

This diff adds a test to `WatchmanWatcher` to ensure that it correctly accounts for `is_fresh_instance`.

While I'm here, I added some tests in which Watchman returns malformed output, and the tests segfaulted. Watchman shouldn't be able to segfault HHVM by changing its JSON, so I changed the parsing code to throw exceptions instead.

Reviewed By: bdchik

Differential Revision: D32272166

fbshipit-source-id: 499df242348328705f5fbb5ab0dd3d00a5b6a338