Refactor HPHP::Watchman to make it mockable
Summary:
This diff converts `HPHP::Watchman` into an abstract class. This allows us to make a mocked version of `HPHP::Watchman` that we can use in tests. Watchman is an IO-heavy service, so being able to mock or fake it out is a huge win towards adding deterministic unit tests to Facts.

See D32272166 for an actual test that uses a mock implementation.

Reviewed By: bdchik

Differential Revision: D32272167

fbshipit-source-id: 106af07fbb065cd9e8a704673a979e736a4027b7