Super inlining hidden behind a flag
Summary:
Super inlining is novel strategy in HHVM attempting to const fold nearly-const
values.  This works by running a custom interpreter (that wraps the reguar
interpreter with hooks) on a set of const or nearly-const values.  The custom
interpreter allows us to use a speciailzed resource type, a MysteryBox, that
represents an unknown value that lives among the known values (perhaps in an
object's property or in a local slot).  Interpretation fails upon inspection of
such a value in an unhandled manner.  The result of super inlining is a self
contained set of heap objects that we copy into a buffer by tracing the heap
pointers.  At runtime we can materialize the runtime representation using an
allocation, memcpy, and several pointer fixups.

Super inlining work on execution traces that conform to two constraints:
- nearly-constness: locals and heap objects are made up of const values, or
  tracked unknown values (eg. the 4th parameter). Of key importance is that
  objects can be considered nearly-const yet they are never truly const
- input-independent: the execution trace's control flow is independent of the
  unknown inputs

Differential Revision: D32259339

fbshipit-source-id: 696dc08cc1e040417f3537e616d5cb17f0d6cdf7