Suppress unsafe cast lint on nondenotable expression type
Summary: The lint was firing on cases where the expression type is not mixed (as expected), but it had a nondenotable type. So `mixed` is in fact a legitimate type that can be used as a source type in this case. This diff suppresses the lint in the described case.

Reviewed By: Wilfred

Differential Revision: D32361202

fbshipit-source-id: 1bc4ffe4970c183723b9aae42316f4cc26cea86d