Increase prof-data file max StringData size
Summary:
Now that we're encoding argument types into the prof-data file, it's
possible to exceed the 2 MB string limit. This is most apparent with
some test cases, which produce giant strings because of const-propping
in the JIT or HHBBC. If they get passed as an argument, they might be
encoded in the prof-data file.

This has always been an issue, as there's nothing preventing us from
having StringData* which are larger than 2 MB. We really should be
setting the limit at StringData::MaxSize instead. "CPP" strings can
continue to have the lower limit.

Reviewed By: oulgen

Differential Revision: D32411999

fbshipit-source-id: c6a9a5e940d09d3c178d67b016152b0b6b91c564