noreturn <D: dynamic
Summary: This diff makes `noreturn <: supportdynamic`. By analogy to void, with a `(function (): void)`, we allow it to upcast to `(function (): dynamic)` since at runtime the implicit return becomes `null`. A `(function (): noreturn)` will never yield control to its caller, so the return value will not used. In that sense, it's similar to `(function (): nothing)`.

Differential Revision: D32424105

fbshipit-source-id: e08950d3efe37a35e0118d886ba879bffa1b20f2