Pull out LdPropAddr into a helper
Summary:
We are going to make it possible to lazily deserialize objects stored in APC. That is, we won't bother localizing an object's properties until they are actually read.

In order to make this work, we need to check that properties are deserialized before accessing them in the JIT. The main way the JIT accesses object properties is via the LdPropAddr / LdInitPropAddr IR ops. Funnel these ops through a central helper, so that we can add these checks when required in this one place.

There are still a couple of uses of LdPropAddr that we don't funnel through this helper - in particular, ones in irgen-bespoke and irgen-create. The basic rule is: we need these checks iff we may read the current value of the property. When constructing a new object, properties are uninitialized, so we can write to them without any checks.

Reviewed By: ricklavoie

Differential Revision: D32435159

fbshipit-source-id: 1648480c6f4b8f02f03b1ff92952723f86408f60