Stop submoduling mcrouter (#8934)
Summary:
This makes mcrouter consistent with the other fbcode dependencies, like folly, thrift, etc, which is generally a move away from custom HHVM-specific CMake code (which was required on CMake 2), to built-in features of CMake 3.

This change also surfaced some nondeterminism in the link order of fizz, so fixed that too.

Added bonuses:
- mcrouter source is no longer downloaded on mac, where we can't build it
- mcrouter source is cached in FB CI (manifold)
- cleaner separation between build directory and source directory means that `git status` is finally empty after a build
- no longer need custom top-level patching system; this can be done via PATCH_COMMAND in ExternalProject_Add if needed in the future (like it is for various other dependencies)
- mcrouter thrift codegen is no longer regenerated every time you run `make`
- cleaner CMake dependency handling for this now means that only the mcrouter HHVM extension depends on mcrouter, rather than all of HHVM

The cumulative effect is that we don't rebuild mcrouter lib (due to it depending on thrift codegen) and lots of core parts of HHVM (due to depending on mcrouter lib) if I type `make hhvm` twice in a row.

Pull Request resolved: https://github.com/facebook/hhvm/pull/8934

Test Plan:
- built on ubuntu 20.04 after deleting system jemalloc.
- mac CI job (which should not build mcrouter)

Reviewed By: shayne-fletcher

Differential Revision: D32395226

fbshipit-source-id: b47f10e7537a04067c50059517082546a044571c