link back to VSCode bugnub
Summary:
When the user clicks the VSCode bug button, it
1. kicks off `hh rage` asynchronously
2. creates a VSCode bugnub
3. a bit later, hh rage finishes and creates another bugnub

It's done this way because VSCode's bug-reporting has a very strict time limit, that `hh rage` exceeds.

Anyway, given a hh-rage bugnub, I'd like to find a way back to the corresponding VSCode bugnub. That's what this diff enables.

WHAT THIS DIFF DOES:

When hh rage is invoked, then we search back for any VSCode bugnubs that have been filed by the same user within the past 10 minutes against the Hack language extension. We include them in an attachment named "VSCode bugs". (Alas the query against flytrap scuba doesn't let us search also by hostname; the only interesting piece of data we can get from it is the title+description).

Also, this diff append the title of that VSCode bugnub to the title we use for our own bug. (I'm hoping that this way, the bug titles in our rage scuba table will be more descriptive).

Reviewed By: madgen

Differential Revision: D32376950

fbshipit-source-id: 4cb2c243adee71ec76e01692891c0486f3c49221