DeclParserOptions includes allow_new_attribute_syntax
Summary: Currently the direct decl parser errors on Attribute syntax, but the facts parser accepts it. `allow_new_attribute_syntax` is the `ParserEnv` flag supporting the syntax. Make `DeclParserOptions` carry the flag `allow_new_attribute_syntax` so we can explicitly set it and have it overwrite the value in `ParserEnv`. Changes in this diff only turn the flag on for decls in comp entrypoints.

Reviewed By: shayne-fletcher

Differential Revision: D32478420

fbshipit-source-id: 5d5aa829e19ed511b4940201185028cab33a0843