More accurate is/as linting
Summary:
The linter check for "always false" or "always true" `is` and `as` expressions misses many "always false" expressions, including any that involve interfaces.

More accurate is to use our new "disjoint types" check: if the type of the scrutinee and the type hint in the `is` or `as` expression are disjoint, then the check will never succeed.

Differential Revision: D32496331

fbshipit-source-id: c677e09a2762ca7a7575690b6be8a8baea4934e9