regression: fix building hack twice
Summary:
We just upgraded the version of dune we use, and due to https://github.com/ocaml/dune/pull/3092,
dune now creates the executables as 555 (r-xr-xr-x) instead of 755.

This is a problem as our later build rules use `cp` to copy the executables into a `bin/` directory. This works once, but on future attempts, it will fail as `cp` will not overwrite read-only files. The dune build is part of a larger,
multi-language project, so we need the end results to fit nicely with other build systems' expectations.

Use `install -m 755` instead.

While I'm here, replace deprecated `opam config exec` with new hotness `opam exec`, as it tells me to do this on every build :)

Reviewed By: vsiles

Differential Revision: D32532812

fbshipit-source-id: 3b87a5e021a7aab61bcc3d6080d2d069d992daa1