Add a HHIR-level peephole optimization to reorder CheckTypes
Summary:
This diff adds a simple pass to optimize two chained, back-to-back CheckType instructions. In general, we should try to avoid emitting such patterns in irgen, but they may appear after optimizations. In particular, they commonly appear after simplifying StructDict accesses, with the first CheckType checking the value is InitCell and the second one checking a specific profiled type, e.g.:

        (6821) t1387:Cell = LdStructDictElem t1386:Dict=StructDict<...>, "..."
        (6822) t1388:InitCell = CheckType<InitCell> t1387:Cell -> B1779<Unused>
        (6823) t1389:Str = CheckType<Str> t1388:InitCell -> B1782<Unlikely>

This pass reorders the CheckTypes to first check for the profiled type, which is expected to succeed, thus saving a check in the common case.

Reviewed By: ricklavoie

Differential Revision: D32412190

fbshipit-source-id: 51faae71b87ed5300f4f0530fd9a90c6b4fce69b