Frame-state and minstr related optimizations
Summary:
Improve frame-state and minstr related operations to make up for the
fact that we don't have pointee types anymore.

Rework frame-state memory handling, particularly around the
mbase. Frame-state can now correctly handle writes to the stack or
locals via StMem. It's tracking is more accurate so it does not need
to pessimize as much. The mTempBase is now a tracked memory location,
as that used commonly in minstr operations, and we want to track the
type of any value stored there.

Frame-state now exposes a function, which given a TMem SSATmp, will
attempt to determine the type stored at the pointee. This is meant to
be used for the situations where we want to know the pointee type
(instead of embedding it into the pointer type). This function will
examine all defs of the SSATmp and use that to determine the type (for
example, if it's def'd by an array operation, it will use
arrLikeElemType). To make this work, some extra type information
needed to be stored in a few HHIR ops.

Frame-state now tracks the mROProp bool as well. Use this tracking in
preOptimize() to remove redundant stores/checks of the bool. Change
LdMROProp to CheckMROProp. This has the advantage that we know the
state of the bool on the next/taken edges.

Utilize this frame-state information in IRBuilder. One interesting
optimization is that we can turn LdMem/StMem/CheckTypeMem/etc into the
equivalent LdLoc/StLoc/CheckLoc (or Stk). In some cases we can
materialize the MBR directly instead of loading it. We did this for a
few special cases in irgen-minstr/irgen-bespoke before, but now it's
completely general.

All these changes allowed for a considerable amount of cleanup in
irgen-minstr and irgen-bespoke.

Reviewed By: jano

Differential Revision: D31954071

fbshipit-source-id: 598c1dc014ea0c7d64e8f203801a3abbe7f3ff91