Remove pointee types from mem lattice in JIT type system
Summary:
The pointee types in the JIT type system are an attractive
nuisance. They're only supposed to be used for types which are
invariant (IE, which don't ever change). An example would be props, or
constants. This property is needed because the pointers are in SSA and
thus cannot ever change type. If you have a TPtrToElemVec, but then
write a Dict to its pointee, the type is now a lie. And, in general,
the JIT is not smart enough to know when a pointer isn't "valid" any
more to use. Frame-state and others will happily re-use the pointer
where you don't expect it.

Unfortunately this restriction isn't always followed. We use pointers
with types in member instruction sequences on things which aren't
technically invariant. We mainly get away with it because the base
mutations aren't sufficiently exposed to the JIT. However, once we try
doing more of this in the TC, we start getting mis-optimizations.

Remove pointee types from the mem lattice of JIT types. Pointer now
only contain what location they point to (IE, we have TPtrToFrame, but
not TPtrToFrameCell). This makes pointers perfectly safe to treat like
any other type, as they are always invariant (a pointer will never
change what location it can point to).

Actual types (where they're needed) are now provided at the point of
usage of the pointer. It is up to irgen and optimization phases to
keep the types as specific as possible. This is the correct approach,
as what type can possibly be contained in a location is a matter for
memory analysis, and not the type-system. Move some simplification
rules and return type determination around into load-elim. In a few
places, we needed to carry around the Type during irgen to provide it
correctly.

As a follow up, I plan on improving analysis in frame-state and
load-elim, which should remove some of the boilerplate from irgen
(particularly during member instructions).

Reviewed By: jano

Differential Revision: D31659422

fbshipit-source-id: 8e1ab773bef88745d93cba3c3c0ce02f94e555ad