Implement more array set operations directly in the TC
Summary:
Instead of performing array set operations solely with C++ helpers,
perform some cases directly in the TC. As long as the set doesn't add
a new key (merely replacing an existing key's value), we can
potentially perform the set in the TC.

We use the existing array profiling machinery for this. Similar to
accesses, we profile the set. If we always get a consistent array
position, we'll attempt to do the set on that position in optimized
translations.

Doing the actual set is easy. We simply obtain a pointer to the
element and then do a StMem with it. However, we need to COW the array
first. We already have a mechanism to see if the array needs to be
COWed. Add a new HHIR op, called CopyArray. This creates a brand new
copy of the array (regardless of the ref-count). So, we check
CheckArrayCOW, and if necessary, we use CopyArray to make a fresh
copy. If we make a copy, we write the new array back into the mbase.

There's an additional complication. Once we start mutating arrays
directly in the TC, we run into problems regarding the
type-system. Mutating the array in place can invalidate SSATmp
types. To keep this from being a problem, we must ensure we never use
a SSATmp of an array beyond the point where it's mutated. CopyArray
already handles this. It's output type drops any specializations
(which may now be invalid), and we write it back to the
base. Frame-state ensures that we won't ever access that old SSATmp
again. However, for the no-copy case, the problem remains. To fix
this, CheckArrayCOW produces a new SSATmp, with the same output type
as CopyArray. We "store" this new SSATmp back to the mbase as well. Of
course, we know that the new SSATmp is actually the same as the
old. So, instead of doing a real store, we use a new "meta" store,
which emits no code. It is solely used to trigger a change in
frame-state (and later memory analysis).

Also add similar support for IncDecM and SetOpM. These operations are
similar to SetM, and we can handle them similarly (for simple
cases). Augment some of the array profiling we're already doing to
provide for more metadata. Namely, not only profile the array offset,
but whether the array will COW, and also perform a dec-ref profile on
the matching value. This lets us emit an efficient dec-ref sequence
when doing the set inline.

Vecs are handled as well. We don't need array access profiling for it,
however, since we can always access the vec element with simple
arithmatic at runtime. Since we can always do this, the ElemVecD and
ElemVecU HHIR ops are now unneeded.

Differential Revision: D32056779

fbshipit-source-id: eae3d0af35a4638002821a2066c4bdb4d925c6a4