Const-fold Vptrs better
Summary:
I noticed a few cases where we weren't const-folding parts of a Vptr
like we would exxpect. We were const folding the base and not the
index.

While doing that, I noticed a few times where vasm-simplify was not
fully const folding cases. I traced it to situations where we were
const folding a Vreg, and then copying the constant to another
Vreg. We didn't really that new Vreg is also a constant, thus weren't
firing the rules. Instead of checking the Vunit's reg-to-const table,
keep our own constant table. The difference is a copy of a constant
Vreg is treated like a constant itself. This lets more folding to
trigger.

Reviewed By: jano

Differential Revision: D31663937

fbshipit-source-id: 2e94a1788e8d6199fb4493bf3c4c9521edbbbcab