Some minstr cleanup
Summary:
Do some cleanup on minstr/pointer related things:

A number of minstr operations on props could take a Lval or a
Cell. However they never modify the base, so they never actually
require a Lval. Just simplify the operations to always take a Cell.

MinstrEffects is useless at this point. There's essentially no
promotion anymore, so the effects are trivial. Any array-likes or
strings just can COW. Just remove MinstrEffects and add the
corresponding logic to frame-state and load-elim. Add a modifier to
type to reflect a particular COW operation.

Rework the TMem hierarchy a bit. A number of the locations were
unused. Remove the dead ones and add a few new useful ones. Rename MIS
to MISTemp to reflect that it only ever points at the minstr temp
base. Add Const to various constant values (like
immutable_null_base). Update IRInstruction src/dsts to reflect the
refined locations more precisely.

Add new AOther alias class to represent misc locations we do not care
about otherwise. It is *not* correct to treat these as AEmpty (as we
were doing). Consider a pointer that could point to a global, or the
minstr temp base. If we map the global to AEmpty, the combined alias
class would just be AMIStateTempBase, and we'd think it would be safe
to replace a load from it with a known type. We need AOther to
represent it could be something else. Add support in memory-effects to
use this properly.

Finally, change LdMIStateAddr to LdMIStateTempBaseAddr since we only
ever used it for the minstr temp base. This means that
PtrToMISTempBase *always* points at the minstr temp base, and nothing
else.

Reviewed By: jano

Differential Revision: D31776520

fbshipit-source-id: e7073cd7b66527f14e132836ddc31cd1c1d4fb6d