Implement a pass to sink defs closer to their usages
Summary:
Sometimes we'll sink stores into side-exits. This leaves SSATmps,
which are now only used by in the side-exit, still on the main trace. This
can be seen most commonly with writes to the MBase. We'll sink the
store to the MBase into a side-exit, but still calculate LdLocAddr or
whatever in the main trace.

Fix this in a general way. Add a pass which attempts to sink instructions
closer to where their SSATmp is used. At best, this will sink a
conditionally needed instruction into a side-exit. Even when it
doesn't do that, it can reduce register lifetimes by shortening SSATmp
live ranges.

The dataflow equations are described in the comments (won't go into
detail here), but it's a fairly trivial PRE-type thing. It has the
advantage that it guarantees a value will never be calculated
redundantly across any path.

For certain instructions, we want to be even more aggressive than
that. For a small set of instructions, we'll rematerialize that
instruction before *every* usage. This is useful for the instructions
which calculate a trivial memory address (like RDS). These addresses
will be folded into their subsequent usage.

Differential Revision: D32336172

fbshipit-source-id: 9381a26cca9ed802b1838df770d9ba9cdc6836d3