Improve type tracking for certain memory locations in load-elim
Summary:
A number of HHIR instructions provide information about the type in a
location (such types are invariant to the location). Right now
load-elim only learns about the types in a location via an explicit
check or store. Have it utilize the information present in these
instructions to "assert" the type in a location. This is similar to
what frame-state is already doing.

Reviewed By: paulbiss

Differential Revision: D32441796

fbshipit-source-id: 9630b837d73eb7d2206108548439fe03ddc38269