Add support for lazy loading of file contents
Summary:
If the file system provides a way to obtain a file's SHA1 without
actually reading it, we can potentially avoid reading the file at
all. We can obtain the file's SHA1, then get a hit with the unit
emitter cache hook. At this point we have everything we need and don't
need to read the file.

Instead of passing the file's contents around, instead use a
LazyUnitContentsLoader. This represents a file's contents, but it
attempts to only load the file's contents only when actually
needed. If the file system does not provide a way to obtain a SHA1, it
will load as eagerly (as normal). Otherwise it will merely obtain the
hash, and wait until the contents are actually requested.

One potential issue here: Since we obtain the file's hash first, then
potentially read the contents later, it's possible for the contents to
change after the hash read. So, when (and if) we read the file's
contents, we calculate the hash from it and check to make sure it
matches the hash we read early. If they disagree, we need to bail. The
caller is generally expected to then retry the operation. In theory
(though it should be extremely rare), this can live-lock. So, if we
detect such inconsistencies a certain number of times, we load the
file's contents eagerly. This shouldn't ever happen.

We support this lazy loading for both sandbox mode and prod builds.

Reviewed By: jano

Differential Revision: D32525102

fbshipit-source-id: e7422f335ace98e613378d32c99bb117d87f0683